added checkout

+++ app/routes.py
@@ -1,8 +1,8 @@
 from flask import render_template, flash, redirect, url_for, request, jsonify
 from app import app, db, mail, photos
 from flask_login import current_user, login_user, login_required, logout_user
-from app.models import User, Company, Submenu, FoodItem
-from app.forms import LoginForm, RegistrationForm, EditProfileForm, ResetPasswordRequestForm, ResetPasswordForm, BecomePartnerForm, EditCompanyForm, AddAdminForm, AddStaffForm, AddSubMenuForm, AddFoodItemForm
+from app.models import User, Company, Submenu, FoodItem, Orders
+from app.forms import LoginForm, RegistrationForm, EditProfileForm, ResetPasswordRequestForm, ResetPasswordForm, BecomePartnerForm, EditCompanyForm, AddAdminForm, AddStaffForm, AddSubMenuForm, AddFoodItemForm, EditFoodItemForm, OrderListForm, OrderCheckoutForm
 from werkzeug.urls import url_parse
 from datetime import datetime
 from app.email import send_password_reset_email
@@ -189,6 +189,8 @@ def user():
 @app.route('/admin', methods=['GET', 'POST'])
 @login_required
 def admin():
+	if not current_user.is_substr_admin():
+		return redirect(url_for('index'))
 	pending_requests = Company.query.filter(Company.date_requested.isnot(None)).filter_by(date_approved=None).all()
 	partners = Company.query.filter(Company.date_approved.isnot(None)).all()
 	return render_template('admin.html', pending_requests=pending_requests, partners=partners)
@@ -197,6 +199,8 @@ def admin():
 @app.route('/approve_request/<comp_id>')
 @login_required
 def approve_request(comp_id):
+	if not current_user.is_substr_admin():
+		return redirect(url_for('index'))
 	approved_company = Company.query.filter_by(id=comp_id).first()
 	approved_company.date_approved = datetime.utcnow()
 	for_hash = approved_company.name + approved_company.date_approved.strftime("%m%d%Y%H%M%S")
@@ -207,6 +211,8 @@ def approve_request(comp_id):
 @app.route('/reject_request/<comp_id>')
 @login_required
 def reject_request(comp_id):
+	if not current_user.is_substr_admin():
+		return redirect(url_for('index'))
 	company = Company.query.get(comp_id)
 	db.session.delete(company)
 	db.session.commit()
@@ -222,12 +228,24 @@ def partner(comp_name, comp_hash):
 	for n in range(0, len(submenu)):
 		menu.show_order = str(n)
 		menu.hash_order = menu.submenu_hash + menu.show_order
-	return render_template('partner.html', company=company, submenu=submenu)
+	orders_form = OrderListForm()
+	if orders_form.validate_on_submit():
+		food_id_list = request.form.getlist('order_food_id')
+		food_hash_list = request.form.getlist('order_food_hash')
+		qty_list = request.form.getlist('order_quantity')
+		for i in range(0, len(food_hash_list)):  #dont set price here so it will be updated for future price change before checkout--set price in checkout
+			list = Orders(food_item_hash=food_hash_list[i], quantity=qty_list[i], food_item_id=food_id_list[i], user_id=current_user.id, company_id=company.id)
+			db.session.add(list)
+			db.session.commit()
+		return redirect(url_for('checkout', username=current_user.username))
+	return render_template('partner.html', company=company, submenu=submenu, orders_form=orders_form)
 	
 @app.route('/manage_company/<comp_name>/<comp_hash>', methods=['GET', 'POST'])
 @login_required
 def manage_company(comp_name, comp_hash):
 	company = Company.query.filter_by(company_hash=comp_hash).first()
+	if not current_user.is_company_admin(company):
+		return redirect(url_for('index'))
 	edit_comp_form = EditCompanyForm(company.email)
 	if edit_comp_form.submit_ecf.data:
 		if edit_comp_form.validate_on_submit():
@@ -259,6 +277,7 @@ def manage_company(comp_name, comp_hash):
 		if add_admin_form.validate_on_submit():
 			user = User.query.filter_by(username=add_admin_form.input_admin.data).first()
 			company.admins.append(user)
+			company.staffs.append(user)
 			db.session.commit()
 			return redirect(url_for('manage_company', comp_hash=company.company_hash, comp_name=company.name))
 	add_staff_form = AddStaffForm()
@@ -284,6 +303,8 @@ def manage_company(comp_name, comp_hash):
 @login_required
 def remove_admin(comp_hash, admin_id):
 	company = Company.query.filter_by(company_hash=comp_hash).first()
+	if not current_user.is_company_admin(company):
+		return redirect(url_for('index'))
 	user = User.query.filter_by(id=admin_id).first()
 	company.admins.remove(user)
 	return redirect(url_for('manage_company', comp_hash=company.company_hash, comp_name=company.name))
@@ -293,6 +314,8 @@ def remove_admin(comp_hash, admin_id):
 @login_required
 def remove_staff(comp_hash, staff_id):
 	company = Company.query.filter_by(company_hash=comp_hash).first()
+	if not current_user.is_company_admin(company):
+		return redirect(url_for('index'))
 	user = User.query.filter_by(id=staff_id).first()
 	company.staffs.remove(user)
 	return redirect(url_for('manage_company', comp_hash=company.company_hash, comp_name=company.name))
@@ -304,6 +327,9 @@ def remove_staff(comp_hash, staff_id):
 def manage_submenu(submenu_name, submenu_hash):
 	submenu = Submenu.query.filter_by(submenu_hash=submenu_hash).first()
 	foods = FoodItem.query.filter_by(submenu_id=submenu.id).order_by(FoodItem.name.asc()).all()
+	company = Company.query.filter_by(id=submenu.company_id).first()
+	if not current_user.is_company_admin(company):
+		return redirect(url_for('index'))
 	edit_submenu_form = AddSubMenuForm()
 	if edit_submenu_form.submit_submenu_form.data:
 		if edit_submenu_form.validate_on_submit():
@@ -314,34 +340,64 @@ def manage_submenu(submenu_name, submenu_hash):
 	food_item_form = AddFoodItemForm()
 	if food_item_form.submit_food_item_form.data:
 		if food_item_form.validate_on_submit():
-			item_query = FoodItem.query.filter_by(name=food_item_form.food_item_name.data, submenu_id=submenu.id).first()
-			if item_query is None:
-				food_item = FoodItem(name=food_item_form.food_item_name.data, price=food_item_form.food_item_price.data, description=food_item_form.food_item_description.data, company_id=submenu.company_id, submenu_id=submenu.id)
-				for_hash = food_item.name + str(food_item.id) + str(submenu.company_id) + str(food_item.price)
-				food_item.set_food_item_hash(for_hash)
-				if food_item_form.food_item_pic.data:
-					pic_query = FoodItem.query.filter_by(cover_pic=str(food_item_form.food_item_pic.data)).first()
-					if pic_query is None:
-						food_item_pic_filename = photos.save(food_item_form.food_item_pic.data)
-						food_item.cover_pic = photos.url(food_item_pic_filename)
-					else:
-						flash('Duplicate cover pic name, please choose a different(specific) name for your cover pic.')
-						return redirect(url_for('manage_submenu', submenu_name=submenu.name, submenu_hash=submenu.submenu_hash))
-				db.session.add(food_item)
-			else:
-				item_query.name = food_item_form.food_item_name.data
-				item_query.price = food_item_form.food_item_price.data
-				item_query.description = food_item_form.food_item_description.data
-				for_hash = item_query.name + str(item_query.id) + str(submenu.company_id) + str(item_query.price)
-				item_query.set_food_item_hash(for_hash)
-				if food_item_form.food_item_pic.data:
-					pic_query = FoodItem.query.filter_by(cover_pic=str(food_item_form.food_item_pic.data)).first()
-					if pic_query is None:
-						food_item_pic_filename = photos.save(food_item_form.food_item_pic.data)
-						item_query.cover_pic = photos.url(food_item_pic_filename)
-					else:
-						flash('Duplicate cover pic name, please choose a different(specific) name for your cover pic.')
-						return redirect(url_for('manage_submenu', submenu_name=submenu.name, submenu_hash=submenu.submenu_hash))
+			food_item = FoodItem(name=food_item_form.food_item_name.data, price=food_item_form.food_item_price.data, description=food_item_form.food_item_description.data, company_id=submenu.company_id, submenu_id=submenu.id)
+			for_hash = food_item.name + str(food_item.id) + str(submenu.company_id) + str(food_item.price)
+			food_item.set_food_item_hash(for_hash)
+			if food_item_form.food_item_pic.data:
+				pic_query = FoodItem.query.filter_by(cover_pic=str(food_item_form.food_item_pic.data)).first()
+				if pic_query is None:
+					food_item_pic_filename = photos.save(food_item_form.food_item_pic.data)
+					food_item.cover_pic = photos.url(food_item_pic_filename)
+				else:
+					flash('Duplicate cover pic name, please choose a different(specific) name for your cover pic.')
+					return redirect(url_for('manage_submenu', submenu_name=submenu.name, submenu_hash=submenu.submenu_hash))
+			db.session.add(food_item)
 			db.session.commit()
 			return redirect(url_for('manage_submenu', submenu_name=submenu.name, submenu_hash=submenu.submenu_hash))
-	return render_template('submenu.html', submenu=submenu, edit_submenu_form=edit_submenu_form, food_item_form=food_item_form, foods=foods)
+	edit_food_item_form = EditFoodItemForm()
+	food_item_list = [(f.id, f.name) for f in foods]
+	edit_food_item_form.edit_food_item_name.choices = food_item_list
+	if edit_food_item_form.validate_on_submit():
+		edit_food_item = FoodItem.query.filter_by(id=edit_food_item_form.edit_food_item_name.data).first()
+		edit_food_item.description = edit_food_item_form.edit_item_description.data
+		if edit_food_item_form.edit_food_item_pic.data:
+				edit_pic_query = FoodItem.query.filter_by(cover_pic=str(edit_food_item_form.edit_food_item_pic.data)).first()
+				if edit_pic_query is None:
+					food_item_pic_filename = photos.save(edit_food_item_form.edit_food_item_pic.data)
+					edit_food_item.cover_pic = photos.url(food_item_pic_filename)
+				else:
+					flash('Duplicate cover pic name, please choose a different(specific) name for your cover pic.')
+					return redirect(url_for('manage_submenu', submenu_name=submenu.name, submenu_hash=submenu.submenu_hash))
+		db.session.commit()
+		return redirect(url_for('manage_submenu', submenu_name=submenu.name, submenu_hash=submenu.submenu_hash))
+	return render_template('submenu.html', submenu=submenu, edit_submenu_form=edit_submenu_form, food_item_form=food_item_form, foods=foods, edit_food_item_form=edit_food_item_form)
+	
+	
+@app.route('/checkout/<username>', methods=['GET', 'POST'])
+@login_required
+def checkout(username):
+	pending_orders = Orders.query.filter_by(user_id=current_user.id, date_purchased=None).all()
+	checkout_form = OrderCheckoutForm()
+	if checkout_form.validate_on_submit():
+		order_id_list = request.form.getlist('order_id')
+		qty_list = request.form.getlist('order_quantity')
+		for i in range(0, len(order_id_list)):
+			order = Orders.query.filter_by(id=order_id_list[i]).first()
+			print(order_id_list[i])
+			order.quantity = qty_list[i]
+			order.price = order.food_item.price  #set price here so that it will be saved and not be affected by future price change
+			order.date_purchased = datetime.utcnow()
+			db.session.commit()
+		return redirect(url_for('index'))
+	return render_template('checkout.html', pending_orders=pending_orders, checkout_form=checkout_form)
+	
+	
+@app.route('/remove_item/<int:id>', methods=['GET', 'POST'])
+@login_required
+def remove_item():
+	item = Order.query.filter_by(id=id).first()
+	if not item.is_ordered_by(current_user):
+		return redirect(url_for('index'))
+	db.session.delete(item)
+	db.session.commit()
+	return redirect(url_for('checkout', username=current_user.username))
